augustcom/backend/requirements.txt
Django>=5.1.4
djangorestframework>=3.16.1
django-filter>=25.1
django-cors-headers>=4.7.0
Pillow>=10.0.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-decouple>=3.8

